[CDEC-416] Move `Infra.Util.{LogSafe,TimeLimit}` to `sinbin`
