+++ block/src/Pos/Block/Error.hs
@@ -9,8 +9,8 @@ module Pos.Block.Error
 
 import           Universum
 
-import           Control.Exception.Safe (Exception (..))
 import           Control.DeepSeq (NFData)
+import           Control.Exception.Safe (Exception (..))
 import           Data.Text.Buildable (Buildable (..))
 import           Data.Text.Lazy.Builder (Builder, fromText)
 import           Formatting (bprint, stext, (%))
