+++ block/src/Pos/Block/Lrc.hs
@@ -38,7 +38,6 @@ import           Pos.DB.Class (MonadDBRead, MonadGState)
 import qualified Pos.DB.GState.Stakes as GS (getRealStake, getRealTotalStake)
 import           Pos.Delegation (getDelegators, isIssuerByAddressHash)
 import qualified Pos.GState.SanityCheck as DB (sanityCheckDB)
-import           Pos.Infra.Util.TimeLimit (logWarningWaitLinear)
 import           Pos.Lrc.Consumer (LrcConsumer (..))
 import           Pos.Lrc.Consumers (allLrcConsumers)
 import           Pos.Lrc.Context (LrcContext (lcLrcSync), LrcSyncData (..))
@@ -53,6 +52,7 @@ import           Pos.Lrc.Types (RichmenStakes)
 import           Pos.Sinbin.Reporting (HasMisbehaviorMetrics (..),
                      MisbehaviorMetrics (..))
 import           Pos.Sinbin.Slotting (MonadSlots)
+import           Pos.Sinbin.Util.TimeLimit (logWarningWaitLinear)
 import           Pos.Ssc (MonadSscMem, noReportNoSecretsForEpoch1,
                      sscCalculateSeed)
 import           Pos.Ssc.Message (SscMessageConstraints)
