+++ ssc/src/Pos/Ssc/Worker.hs
@@ -37,13 +37,13 @@ import           Pos.Infra.Binary ()
 import           Pos.Infra.Diffusion.Types (Diffusion (..))
 import           Pos.Infra.Recovery.Info (recoveryCommGuard)
 import           Pos.Infra.Slotting (getSlotStartEmpatically, onNewSlot)
-import           Pos.Infra.Util.LogSafe (logDebugS, logErrorS, logInfoS,
-                     logWarningS)
 import           Pos.Lrc.Consumer.Ssc (getSscRichmen)
 import           Pos.Lrc.Types (RichmenStakes)
 import           Pos.Sinbin.Reporting (HasMisbehaviorMetrics (..),
                      MisbehaviorMetrics (..))
 import           Pos.Sinbin.Slotting (defaultOnNewSlotParams, getCurrentSlot)
+import           Pos.Sinbin.Util.LogSafe (logDebugS, logErrorS, logInfoS,
+                     logWarningS)
 import           Pos.Ssc.Base (isCommitmentIdx, isOpeningIdx, isSharesIdx,
                      mkSignedCommitment)
 import           Pos.Ssc.Behavior (SscBehavior (..), SscOpeningParams (..),
